requirements.txt
Django==5.2.4
djangorestframework==3.16.0
psycopg2-binary==2.9.10
django-phonenumber-field==8.1.0
phonenumbers==9.0.10
drf-spectacular==0.28.0